move steps components one dir up
