+++ src/shared/components/guidedTour/steps/index.ts
@@ -1,5 +1,5 @@
-export { AreasStep } from './areas/areasStep.component';
-export { InitialStep } from './initial/initialStep.component';
-export { SkillsStep } from './skills/skillsStep.component';
-export { TechnologyPopupStep } from './technologyPopup/technologyPopupStep.component';
-export { TeamsStep } from './teams/teamsStep.component';
+export { AreasStep } from './areasStep.component';
+export { InitialStep } from './initialStep.component';
+export { SkillsStep } from './skillsStep.component';
+export { TechnologyPopupStep } from './technologyPopupStep.component';
+export { TeamsStep } from './teamsStep.component';
